requirements.txt
Flask==2.2.3
pytesseract==0.3.9
openai==0.27.0
numpy==1.23.5
Pillow==9.4.0
python-dotenv==0.21.0
